+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0
+
+- Adds support for shell commands via the `$()` syntax
+- new documentation pages added to introduce generators and more examples
+- Deletes deprecated `Dotenvy.env/3` function
+- Updates dependencies to latest
+
 ## v0.9.0
 
 - Removes fallback to `System.fetch_env/1` and instead requires explicit sourcing of system envs. See [Issue 21](https://github.com/fireproofsocks/dotenvy/issues/21)